Otimizei o projeto.
Agora ele calcula fibonaccis maiores, e o makefile
está mais completo, gerando e mantendo os arquivos objeto, facilitando
as recompições futuras

+++ src/fib.h
@@ -1,7 +1,7 @@
 #ifndef FIB_H
 #define FIB_H
 
-int fib_recursivo(int);
-int fib_iterativo(int);
+unsigned long long int fib_recursivo(int);
+unsigned long long int fib_iterativo(int);
 
 #endif // FIB_H
